change some int that represent Java's long to int64 (for #83)

+++ get_next_operation_id_command.go
@@ -9,7 +9,7 @@ var (
 )
 
 type _GetNextOperationIDCommandResponse struct {
-	ID int `json:"Id"`
+	ID int64 `json:"Id"`
 }
 
 // GetNextOperationIDCommand represents command for getting next
@@ -17,7 +17,7 @@ type _GetNextOperationIDCommandResponse struct {
 type GetNextOperationIDCommand struct {
 	RavenCommandBase
 
-	Result int
+	Result int64
 }
 
 // NewGetNextOperationIDCommand returns GetNextOperationIDCommand
