+++ put_connection_string_result.go
@@ -2,5 +2,5 @@ package ravendb
 
 // PutConnectionStringResult describes result of "put connection" command
 type PutConnectionStringResult struct {
-	Etag int `json:"ETag"`
+	Etag int64 `json:"ETag"`
 }
